@@ -1,34 +1,29 @@
-import Vue from 'vue';
-import Router from 'vue-router';
-import Home from '@/views/Home.vue';
-import Quiz from '@/components/Quiz.vue';
-import NotFound from '@/views/NotFound.vue';
-Vue.use(Router);
+import { createRouter, createWebHistory } from 'vue-router'
+import HomeView from '../views/HomeView.vue'
+import Quiz from '@/components/Quiz.vue'
 
-const router = new Router({
-	mode: 'history',
-	base: process.env.BASE_URL,
-	routes: [
-		{
-			path: '/',
-			name: 'home',
-			component: Home,
-		},
-		{
-			path: '/quiz/:id',
-			name: 'Quiz',
-			component: Quiz,
-		},
-		{
-			path: '/:pathMatch(.*)*',
-			name: 'NotFound',
-			component: NotFound,
-			meta: { title: 'Not Found' },
-		},
-	],
-});
+const router = createRouter({
+  history: createWebHistory(import.meta.env.BASE_URL),
+  routes: [
+    {
+      path: '/',
+      name: 'home',
+      component: HomeView
+    },
+    {
+      path: '/quiz/:id',
+      name: 'quiz-detail',
+      component: Quiz
+    },
+    {
+      path: '/about',
+      name: 'about',
+      // route level code-splitting
+      // this generates a separate chunk (About.[hash].js) for this route
+      // which is lazy-loaded when the route is visited.
+      component: () => import('../views/AboutView.vue')
+    }
+  ]
+})
 
-router.beforeEach((to, from, next) => {
-	next();
-});
-export default router;
+export default router
